deda-ingestor/pyproject.toml
[tool.poetry]
name = "deda-ingestor"
version = "0.1.0"
description = "A Python application that processes product data from RabbitMQ and syncs it with Ivanti"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "deda_ingestor", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pika = "^1.3.2"           # RabbitMQ client
httpx = "^0.26.0"         # Modern HTTP client
pydantic = "^2.5.3"       # Data validation
python-dotenv = "^1.0.0"  # Environment variables management
apscheduler = "^3.10.4"   # Job scheduling
tenacity = "^8.2.3"       # Retry mechanism
structlog = "^24.1.0"     # Structured logging
dependency-injector = "^4.41.0"  # Dependency injection container

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.3"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=deda_ingestor --cov-report=term-missing"
asyncio_mode = "auto"
